* new: color::set::* * new: color::get::gray * fix: set want work for uint64_t/uint32_t * fix: gray<uint16_t>, gray<uint32_t>, gray<uint64_t>, gray<uint8_t> container/component was wrang * fix: rgb<uint32_t>, gray<uint64_t> container/component was wrang * new: tests for get/set * improvement: better output of tests

+++ example/print.hpp
@@ -11,7 +11,7 @@ template< typename category_name >
 
    for( typename model_type::index_type index=0; index < model_type::size(); ++index )
     {
-     std::cout << " " << gray.get( index ) << ", ";
+     std::cout << " " << double( gray.get( index ) ) << ", ";
     }
     std::cout << ">";
   }
