+++ src/gray/uint32.hpp
@@ -24,7 +24,7 @@ namespace color
 
     template< >
      struct container< ::color::category::gray_uint32 >
-      : public  ::color::_internal::utility::container::Unsigned< std::uint32_t, std::uint8_t,  unsigned,  1, 32  >
+      : public  ::color::_internal::utility::container::Unsigned< std::uint32_t, std::uint32_t, unsigned, 1, 32  >
       {
       };
